refactor(modal): extract owner write helper boundary
- Extract owner-write helper functions into new modal_compute/owner_writes.py
- Keep route decorators and route handlers in modal_compute/app.py
- Update imports to use owner_writes module
- Update contract tests to read from owner_writes.py

Behavior preserved:
- Auth, owner boundary, validation, response shape unchanged
- DB queries and error handling unchanged
- Route behavior unchanged

Refs #555
Depends-on/completed: #554
Notes: #553 owner-read extraction already completed and preserved
